Move read_as() to core module
