+++ src/lgdo/lh5/__init__.py
@@ -6,10 +6,10 @@ browsed easily in python like any `HDF5 <https://www.hdfgroup.org>`_ file using
 """
 from __future__ import annotations
 
-from .core import read
+from .core import read, read_as
 from .iterator import LH5Iterator
 from .store import LH5Store
-from .tools import load_dfs, load_nda, ls, read_as, show
+from .tools import load_dfs, load_nda, ls, show
 from .utils import read_n_rows
 
 __all__ = [
