+++ src/lgdo/lh5/core.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
+import inspect
 import sys
 from collections.abc import Mapping, Sequence
+from typing import Any
 
 import h5py
 from numpy.typing import ArrayLike
@@ -119,3 +121,48 @@ def read(
     )
 
     return obj if obj_buf is None else (obj, n_rows_read)
+
+
+def read_as(
+    name: str,
+    lh5_file: str | h5py.File | Sequence[str | h5py.File],
+    library: str,
+    **kwargs,
+) -> Any:
+    """Read LH5 data from disk straight into a third-party data format view.
+
+    This function is nothing more than a shortcut chained call to
+    :meth:`.LH5Store.read` and to :meth:`.LGDO.view_as`.
+
+    Parameters
+    ----------
+    name
+        LH5 object name on disk.
+    lh5_file
+        LH5 file name.
+    library
+        string ID of the third-party data format library (``np``, ``pd``,
+        ``ak``, etc).
+
+    See Also
+    --------
+    .read, .LGDO.view_as
+    """
+    # determine which keyword arguments should be forwarded to read() and which
+    # should be forwarded to view_as()
+    read_kwargs = inspect.signature(read).parameters.keys()
+
+    kwargs1 = {}
+    kwargs2 = {}
+    for k, v in kwargs.items():
+        if k in read_kwargs:
+            kwargs1[k] = v
+        else:
+            kwargs2[k] = v
+
+    # read the LGDO from disk
+    # NOTE: providing a buffer does not make much sense
+    obj = read(name, lh5_file, **kwargs1)
+
+    # and finally return a view
+    return obj.view_as(library, **kwargs2)
